TIDIJ-199 Tested function getShortWeekday and started code optimization.
