+++ fleetcarpooling/lib/utils/datetime_utils.dart
@@ -1,0 +1,22 @@
+class DateTimeUtils {
+  String getShortWeekday(DateTime dateTime) {
+    switch (dateTime.weekday) {
+      case DateTime.monday:
+        return 'Mon';
+      case DateTime.tuesday:
+        return 'Tue';
+      case DateTime.wednesday:
+        return 'Wed';
+      case DateTime.thursday:
+        return 'Thu';
+      case DateTime.friday:
+        return 'Fri';
+      case DateTime.saturday:
+        return 'Sat';
+      case DateTime.sunday:
+        return 'Sun';
+      default:
+        return '';
+    }
+  }
+}
